Fixed some collision problems
Fixed enemy's troops walking too far into playerr's castle, before stopping - problem was that I added enemy troop's reach (which is negative) resulting in the area they stop in to be smaller.
Also added some boundries to troops, which are used for when there needs to be a little room between allies, but not 'reach'-lengths of room.